+++ osu.Framework/osu.Framework.csproj
@@ -310,7 +310,6 @@
     <Compile Include="Allocation\DependencyContainer.cs" />
     <Compile Include="Allocation\BackgroundDependencyLoader.cs" />
     <Compile Include="Configuration\BindableNumber.cs" />
-    <Compile Include="Graphics\UserInterface\CircularProgress.cs" />
     <Compile Include="Graphics\UserInterface\CircularProgressDrawNode.cs" />
     <Compile Include="Graphics\UserInterface\CircularProgress2.cs" />
   </ItemGroup>
